Complete
Could be interesting to explore more on how your code breaks ties in the event that several policies are equally optimal. This tie breaking currently just selects the first index that is best, so that results in jumps in your policy function. In the future, you could figure out a way to make is smooth, ex by referencing the optimal policy of the previous (and next) state and seeing which is best.