Seems to work
Has some quirks, for example, doesn't print policy if under 400 episodes or so (that seems to be the around minimum number of episodes for it to find an optimal policy. Unclear to me why it doesn't work below that threshold). Also, Chatgpt wrote the graphing portion, so I am not sure exactly how that all works.